For #172: Review corrections
