+++ timeless/__init__.py
@@ -90,3 +90,5 @@ def register_endpoints(app):
     )
     app.register_blueprint(floors_views.bp)
     app.register_blueprint(table_shapes_views.bp)
+    app.register_blueprint(reservations_views.bp)
+
