--- timeless/restaurants/reservations/views.py
@@ -1,46 +1,0 @@
-"""Reservations views module.
-
-"""
-
-
-from flask import (
-    Blueprint, flash, redirect, render_template, request, url_for
-)
-
-
-bp = Blueprint("reservations", __name__, url_prefix="/reservations")
-
-
-@bp.route("/")
-def list_reservations(reservations):
-    flash("List not yet implemented")
-    return render_template(
-        "restaurants/tables/list.html", reservations=reservations
-    )
-
-
-@bp.route("/create", methods=("GET", "POST"))
-def create(reservation):
-    if request.method == "POST":
-        flash("Create not yet implemented")
-
-    return render_template(
-        "restaurants/tables/create_edit.html", action="create",
-        reservation=reservation
-    )
-
-
-@bp.route("/edit/<int:id>", methods=("GET", "POST"))
-def edit(id):
-    if request.method == "POST":
-        flash("Edit not yet implemented")
-    return render_template(
-        "restaurants/tables/create_edit.html", action="edit",
-        id=id
-    )
-
-
-@bp.route("/delete", methods=["POST"])
-def delete():
-    flash("Delete not yet implemented")
-    return redirect(url_for("reservations.list_reservations"))
